requirements.txt
streamlit==1.44.1
anthropic==0.83.0
chromadb==1.5.1
pdfplumber==0.11.9
rank-bm25==0.2.2
python-dotenv==1.1.0
cohere==5.20.6
